+++ Rakefile
@@ -5,6 +5,8 @@ task :test do |t|
   targets = [
     "test/rgot_common_test.rb",
     "test/rgot_test.rb",
+    "test/rgot_benchmark_test.rb",
+    "test/rgot_example_test.rb",
   ]
   ruby "bin/rgot -v #{targets.join(' ')}"
 end
